a pair of functions that cache the inverse of a matrix
The first function helps in creating matrix which then sets it into list
containing the functions to set the matrix, get the matrix, set the
inverse of the matrix and get the inverse of the matrix.
The second function checks if the inverse has already been calculated.
If not, it calculates the inverse and gets the answer otherwise it
retrieves the already calculated inverse